@@ -4,11 +4,14 @@
     "minimum-stability": "stable",
     "prefer-stable": true,
     "require": {
-        "php": ">=7.2.5",
+        "php": ">=8.1",
         "ext-ctype": "*",
         "ext-iconv": "*",
+        "ext-openssl": "*",
         "doctrine/annotations": "^2.0",
-        "sensio/framework-extra-bundle": "^6.2",
+        "doctrine/doctrine-bundle": "^2.8",
+        "doctrine/doctrine-migrations-bundle": "^3.2",
+        "doctrine/orm": "^2.14",
         "symfony/apache-pack": "^1.0",
         "symfony/asset": "5.4.*",
         "symfony/console": "5.4.*",
@@ -26,7 +29,9 @@
         "symfony/validator": "5.4.*",
         "symfony/webpack-encore-bundle": "^1.16",
         "symfony/yaml": "5.4.*",
-        "twig/extra-bundle": "^2.12|^3.0",
+        "symfonycasts/verify-email-bundle": "^1.13",
+        "twig/cssinliner-extra": "^3.5",
+        "twig/extra-bundle": "^3.5",
         "twig/twig": "^2.12|^3.0"
     },
     "config": {
