@@ -2,23 +2,39 @@
 
 namespace App\Service;
 
+use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 
 class APIService {
 
-    private $api;
-
-    public function __construct(HttpClientInterface $httpClient) {
+    public function __construct(HttpClientInterface $httpClient, private readonly ContainerInterface $container) {
         $this->httpClient = $httpClient->withOptions([
         'headers' => ['Accept' => 'application/json']
         ]);
+    }
+
+    public function decryptData($data) : array
+    {
+        $path= 'file://'.$this->container->getParameter('kernel.project_dir').'/config/keys/private.pem';
+        $privateKey=openssl_pkey_get_private($path, 'geronimo');
+        $encryptedData = base64_decode($data['value']);
+        $blockSize = 256;
+        $blocks = str_split($encryptedData, $blockSize);
+        $decryptedBlocks = [];
+        foreach ($blocks as $block) {
+            $decryptedBlock = '';
+            openssl_private_decrypt($block, $decryptedBlock, $privateKey);
+            $decryptedBlocks[] = $decryptedBlock;
+        }
+        $decryptedData = implode('', $decryptedBlocks);
+        return json_decode($decryptedData, true);
     }
 
     public function getClub(): array {
         try {
             $response = $this->httpClient->request(
                     'GET',
-                    'http://api/api/clubs?page=1'
+                    'http://api/api/clubs/'
                     
             );
 
@@ -32,10 +48,11 @@ class APIService {
         try {
             $response = $this->httpClient->request(
                     'GET',
-                    'http://api/api/licencies?page=1'
+//                    'http://api/api/licencies/'
+                'http://php-symfony-api:80/api/licencies/',
             );
 
-            return $response->toArray();
+            return $this->decryptData($response->toArray());;
         } catch (Exception $ex) {
             
         }
@@ -53,5 +70,21 @@ class APIService {
             
         }
     }
+
+    public function getLicencieById(int $id): ?array
+    {
+            $response = $this->httpClient->request(
+                'GET',
+                'http://php-symfony-api:80/api/licencies/'.$id,
+//                'http://api/api/licencies/'.$id,
+            );
+            if($response->getStatusCode() === 200)
+            {
+                return $this->decryptData($response->toArray());
+            }else {
+                return null;
+            }
+    }
+
 
 }
