@@ -6,7 +6,11 @@
  */
 
 // any CSS you import will output into a single css file (app.css in this case)
-import './styles/app.css';
+import './styles/app.scss';
+
 
 // start the Stimulus application
 import './bootstrap';
+import './scripts/materialize-init';
+import './scripts/register';
+import './scripts/modal';
