@@ -7,70 +7,81 @@ if (!Encore.isRuntimeEnvironmentConfigured()) {
 }
 
 Encore
-    // directory where compiled assets will be stored
-    .setOutputPath('public/build/')
-    // public path used by the web server to access the output path
-    .setPublicPath('/build')
-    // only needed for CDN's or subdirectory deploy
-    //.setManifestKeyPrefix('build/')
+        // directory where compiled assets will be stored
+        .setOutputPath('public/build/')
+        .copyFiles({
+            from: './assets/image',
+            // optional target path, relative to the output dir
+            to: 'image/[path][name].[ext]',
 
-    /*
-     * ENTRY CONFIG
-     *
-     * Each entry will result in one JavaScript file (e.g. app.js)
-     * and one CSS file (e.g. app.css) if your JavaScript imports CSS.
-     */
-    .addEntry('app', './assets/app.js')
+            // if versioning is enabled, add the file hash too
+            //to: 'images/[path][name].[hash:8].[ext]',
 
-    // enables the Symfony UX Stimulus bridge (used in assets/bootstrap.js)
-    .enableStimulusBridge('./assets/controllers.json')
+            // only copy files matching this pattern
+            //pattern: /.(png|jpg|jpeg)$/
+        })
+        // public path used by the web server to access the output path
+        .setPublicPath('/build')
+        // only needed for CDN's or subdirectory deploy
+        //.setManifestKeyPrefix('build/')
 
-    // When enabled, Webpack "splits" your files into smaller pieces for greater optimization.
-    .splitEntryChunks()
+        /*
+         * ENTRY CONFIG
+         *
+         * Each entry will result in one JavaScript file (e.g. app.js)
+         * and one CSS file (e.g. app.css) if your JavaScript imports CSS.
+         */
+        .addEntry('app', './assets/app.js')
 
-    // will require an extra script tag for runtime.js
-    // but, you probably want this, unless you're building a single-page app
-    .enableSingleRuntimeChunk()
+        // enables the Symfony UX Stimulus bridge (used in assets/bootstrap.js)
+        .enableStimulusBridge('./assets/controllers.json')
 
-    /*
-     * FEATURE CONFIG
-     *
-     * Enable & configure other features below. For a full
-     * list of features, see:
-     * https://symfony.com/doc/current/frontend.html#adding-more-features
-     */
-    .cleanupOutputBeforeBuild()
-    .enableBuildNotifications()
-    .enableSourceMaps(!Encore.isProduction())
-    // enables hashed filenames (e.g. app.abc123.css)
-    .enableVersioning(Encore.isProduction())
+        // When enabled, Webpack "splits" your files into smaller pieces for greater optimization.
+        .splitEntryChunks()
 
-    // configure Babel
-    // .configureBabel((config) => {
-    //     config.plugins.push('@babel/a-babel-plugin');
-    // })
+        // will require an extra script tag for runtime.js
+        // but, you probably want this, unless you're building a single-page app
+        .enableSingleRuntimeChunk()
 
-    // enables and configure @babel/preset-env polyfills
-    .configureBabelPresetEnv((config) => {
-        config.useBuiltIns = 'usage';
-        config.corejs = '3.23';
-    })
+        /*
+         * FEATURE CONFIG
+         *
+         * Enable & configure other features below. For a full
+         * list of features, see:
+         * https://symfony.com/doc/current/frontend.html#adding-more-features
+         */
+        .cleanupOutputBeforeBuild()
+        .enableBuildNotifications()
+        .enableSourceMaps(!Encore.isProduction())
+        // enables hashed filenames (e.g. app.abc123.css)
+        .enableVersioning(Encore.isProduction())
 
-    // enables Sass/SCSS support
-    //.enableSassLoader()
+        // configure Babel
+        // .configureBabel((config) => {
+        //     config.plugins.push('@babel/a-babel-plugin');
+        // })
 
-    // uncomment if you use TypeScript
-    //.enableTypeScriptLoader()
+        // enables and configure @babel/preset-env polyfills
+        .configureBabelPresetEnv((config) => {
+            config.useBuiltIns = 'usage';
+            config.corejs = '3.23';
+        })
 
-    // uncomment if you use React
-    //.enableReactPreset()
+        // enables Sass/SCSS support
+        .enableSassLoader()
 
-    // uncomment to get integrity="..." attributes on your script & link tags
-    // requires WebpackEncoreBundle 1.4 or higher
-    //.enableIntegrityHashes(Encore.isProduction())
+        // uncomment if you use TypeScript
+        //.enableTypeScriptLoader()
 
-    // uncomment if you're having problems with a jQuery plugin
-    //.autoProvidejQuery()
-;
+        // uncomment if you use React
+        //.enableReactPreset()
+
+        // uncomment to get integrity="..." attributes on your script & link tags
+        // requires WebpackEncoreBundle 1.4 or higher
+        //.enableIntegrityHashes(Encore.isProduction())
+
+        // uncomment if you're having problems with a jQuery plugin
+        //.autoProvidejQuery()
+        ;
 
 module.exports = Encore.getWebpackConfig();
